+++ src/main/resources/mediator.properties
@@ -5,7 +5,7 @@ mediator.port=3000
 mediator.timeout=60000
 mediator.heartbeats=true
 
-core.host=localhost
+core.host=139.162.149.249
 core.api.port=8080
 core.api.user=root@openhim.org
-core.api.password=openhim-password
+core.api.password=HimTZ2020.
